refactor: also add a `paths` module
