+++ crates/alembic-engine/src/paths.rs
@@ -1,0 +1,182 @@
+use anyhow::{anyhow, Result};
+use serde_yaml::Value as YamlValue;
+
+#[derive(Debug, Clone)]
+pub(crate) enum SelectorToken {
+    Key(String),
+    Index(usize),
+    Wildcard,
+}
+
+#[derive(Debug, Clone)]
+pub(crate) enum PathToken {
+    Key(String),
+    Index(usize),
+}
+
+#[derive(Debug)]
+pub(crate) struct RelativePath {
+    up: usize,
+    selectors: Vec<SelectorToken>,
+}
+
+pub(crate) fn parse_selector_path(path: &str) -> Result<Vec<SelectorToken>> {
+    if !path.starts_with('/') {
+        return Err(anyhow!("select path must start with '/'"));
+    }
+    let mut tokens = Vec::new();
+    for segment in path.trim_start_matches('/').split('/') {
+        if segment.is_empty() {
+            continue;
+        }
+        tokens.push(parse_selector_segment(segment)?);
+    }
+    Ok(tokens)
+}
+
+fn parse_selector_segment(segment: &str) -> Result<SelectorToken> {
+    if segment == "*" {
+        return Ok(SelectorToken::Wildcard);
+    }
+    if let Ok(index) = segment.parse::<usize>() {
+        return Ok(SelectorToken::Index(index));
+    }
+    Ok(SelectorToken::Key(segment.to_string()))
+}
+
+pub(crate) fn parse_relative_path(path: &str) -> Result<RelativePath> {
+    let mut rest = path.trim();
+    let mut up = 0;
+    while rest.starts_with('^') {
+        up += 1;
+        rest = &rest[1..];
+        if rest.starts_with('.') {
+            rest = &rest[1..];
+        }
+    }
+    if rest.starts_with('.') {
+        rest = &rest[1..];
+    }
+    if rest.starts_with('/') {
+        rest = &rest[1..];
+    }
+    let selectors = if rest.is_empty() {
+        Vec::new()
+    } else {
+        rest.split('/')
+            .filter(|s| !s.is_empty())
+            .map(parse_selector_segment)
+            .collect::<Result<Vec<_>>>()?
+    };
+    Ok(RelativePath { up, selectors })
+}
+
+pub(crate) fn extract_values<'a>(
+    raw: &'a YamlValue,
+    path: &[PathToken],
+    rel: &RelativePath,
+) -> Result<Vec<&'a YamlValue>> {
+    let base_path = ancestor_path(raw, path, rel.up)?;
+    let Some(base_value) = value_at_path(raw, &base_path) else {
+        return Ok(Vec::new());
+    };
+    let mut results = Vec::new();
+    select_values(base_value, &rel.selectors, &mut results);
+    Ok(results)
+}
+
+fn ancestor_path(raw: &YamlValue, path: &[PathToken], up: usize) -> Result<Vec<PathToken>> {
+    let mut current: Vec<PathToken> = path.to_vec();
+    for _ in 0..up {
+        if current.is_empty() {
+            return Err(anyhow!("relative path escapes above root"));
+        }
+        current.pop();
+        while let Some(value) = value_at_path(raw, &current) {
+            if matches!(value, YamlValue::Sequence(_)) {
+                if current.is_empty() {
+                    break;
+                }
+                current.pop();
+            } else {
+                break;
+            }
+        }
+    }
+    Ok(current)
+}
+
+fn value_at_path<'a>(value: &'a YamlValue, path: &[PathToken]) -> Option<&'a YamlValue> {
+    let mut current = value;
+    for token in path {
+        match token {
+            PathToken::Key(key) => {
+                let YamlValue::Mapping(map) = current else {
+                    return None;
+                };
+                current = map.get(YamlValue::String(key.clone()))?;
+            }
+            PathToken::Index(index) => {
+                let YamlValue::Sequence(items) = current else {
+                    return None;
+                };
+                current = items.get(*index)?;
+            }
+        }
+    }
+    Some(current)
+}
+
+fn select_values<'a>(
+    value: &'a YamlValue,
+    selectors: &[SelectorToken],
+    results: &mut Vec<&'a YamlValue>,
+) {
+    if selectors.is_empty() {
+        results.push(value);
+        return;
+    }
+    match selectors[0].clone() {
+        SelectorToken::Key(key) => {
+            if let YamlValue::Mapping(map) = value {
+                if let Some(value) = map.get(YamlValue::String(key)) {
+                    select_values(value, &selectors[1..], results);
+                }
+            }
+        }
+        SelectorToken::Index(index) => {
+            if let YamlValue::Sequence(items) = value {
+                if let Some(value) = items.get(index) {
+                    select_values(value, &selectors[1..], results);
+                }
+            }
+        }
+        SelectorToken::Wildcard => match value {
+            YamlValue::Sequence(items) => {
+                for value in items {
+                    select_values(value, &selectors[1..], results);
+                }
+            }
+            YamlValue::Mapping(map) => {
+                for (key, value) in map {
+                    if key.as_str().is_none() {
+                        continue;
+                    }
+                    select_values(value, &selectors[1..], results);
+                }
+            }
+            _ => {}
+        },
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    #[test]
+    fn parse_relative_path_tracks_parent_hops() {
+        let rel = parse_relative_path("^^.slug").unwrap();
+        assert_eq!(rel.up, 2);
+        assert_eq!(rel.selectors.len(), 1);
+    }
+}
